+++ src/models/lstm.py
@@ -1,61 +1,32 @@
 from tensorflow.keras import layers, Model
 
 
-class LSTMBottleneck(layers.Layer):
-    """https://stackoverflow.com/a/59313862/1834892"""
-
-    def __init__(self, lstm_units, time_steps, **kwargs):
-        self.lstm_units = lstm_units
-        self.time_steps = time_steps
-        self.lstm_layer = layers.LSTM(lstm_units, return_sequences=False)
-        self.repeat_layer = layers.RepeatVector(time_steps)
-        super().__init__(**kwargs)
-
-    def call(self, inputs):
-        # just call the two initialized layers
-        return self.repeat_layer(self.lstm_layer(inputs))
-
-    def compute_mask(self, inputs, mask=None):
-        # return the input_mask directly
-        return mask
-
-
 class LSTMAutoEncoder:
+    """An LSTM Autoencoder model. Later this can be expanded to include prediction,
+    clustering, and reidentification outputs."""
+
     def __init__(
         self,
         optimizer,
         loss,
         n_timesteps: int,
         n_feature_values: int,
-        embedding_length: int,
         lstm_units: int,
-        activation="tanh",
         metrics=None,
     ):
-        """"""
-        inputs = layers.Input(shape=(n_timesteps,))
-        embedding = layers.Embedding(
-            n_feature_values,
-            embedding_length,
-            input_length=n_timesteps,
-            mask_zero=True,
-        )(inputs)
-        # masking = layers.Masking(mask_value=mask_value)(inputs)
-        # encoder = LSTMBottleneck(lstm_units, n_timesteps)(masking)
-        # encoder = LSTMBottleneck(
-
-        encoder = layers.LSTM(lstm_units, input_shape=(n_timesteps, 1))(
-            embedding
-        )
-        hidden = layers.RepeatVector(n_timesteps)(encoder)
-        decoder = layers.LSTM(
-            lstm_units,
-            activation=activation,
+        """Constructor"""
+        inputs = layers.Input(shape=(n_timesteps, 1))
+        encoder_1 = layers.LSTM(lstm_units, return_sequences=True)(inputs)
+        encoder_2 = layers.LSTM(lstm_units // 2)(encoder_1)
+        hidden = layers.RepeatVector(n_timesteps)(encoder_2)
+        decoder_1 = layers.LSTM(
+            lstm_units // 2,
             return_sequences=True,
         )(hidden)
+        decoder_2 = layers.LSTM(lstm_units, return_sequences=True)(decoder_1)
         outputs = layers.TimeDistributed(
             layers.Dense(n_feature_values, activation="softmax")
-        )(decoder)
+        )(decoder_2)
         encoder_model = Model(inputs=inputs, outputs=outputs)
         encoder_model.compile(optimizer=optimizer, loss=loss, metrics=metrics)
         self.model = encoder_model
@@ -82,3 +53,6 @@ class LSTMAutoEncoder:
             callbacks=callbacks,
             verbose=verbose,
         )
+
+    def predict(self, x):
+        return self.model.predict(x)
