+++ src/models/preprocessing.py
@@ -1,6 +1,11 @@
 from pathlib import Path
 import numpy as np
 import pandas as pd
+from src.lib import config
+from src.models.lstm import LSTMAutoEncoder
+from tensorflow.keras.preprocessing.sequence import pad_sequences
+from sklearn.preprocessing import OrdinalEncoder
+
 
 from src.lib import config, freq_matrix, preprocess
 
@@ -35,7 +40,8 @@ def get_trajectory_samples(
     end_date: pd.Timestamp,
     sample_resolution: str,
     trajectory_period: str,
-    pad_value: float,
+    min_points_per_trajectory: int,
+    pad_value: float = 0.0,
 ) -> pd.DataFrame:
     """
     Parameters
@@ -52,6 +58,8 @@ def get_trajectory_samples(
     trajectory_period: str
         A pandas DateOffset string indicating the length of one trajectory example in time.
         See https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html#dateoffset-objects
+    min_points_per_trajectory: int
+        User trajectories with fewer points than this will get filtered out.
     pad_value: float
         The value to fill / pad missing periods in the time series.
     """
@@ -82,8 +90,112 @@ def get_trajectory_samples(
     df["Period"] = (
         pd.DatetimeIndex(df["DateTime"]).to_period(trajectory_period).start_time
     )
-    # exclude periods where the user didn't have at least 3 points
+    # exclude periods where the user didn't have enough points
     df = df.groupby(["UID", "Period"]).filter(
-        lambda x: np.count_nonzero(x["Position"]) >= 3
+        lambda x: np.count_nonzero(x["Position"]) >= min_points_per_trajectory
     )
     return df
+
+
+def preprocess_time_gps(df, start_date, end_date, sample_resolution, pad_value=0):
+    """In case we need to train on raw GPS later."""
+    df["DateTime"] = pd.to_datetime(df["Date"] + " " + df["Time"])
+    df = df[(df["DateTime"] >= start_date) & (df["DateTime"] < end_date)]
+    expanded_range = pd.date_range(
+        start=start_date,
+        end=end_date,
+        closed="left",
+        freq=sample_resolution,
+        name="DateTime",
+    )
+    df = (
+        (
+            df[["DateTime", "UID", "Latitude", "Longitude"]]
+            .groupby("UID")
+            .apply(
+                lambda g: g.set_index("DateTime")
+                .resample(sample_resolution)
+                .mean()[["Latitude", "Longitude"]]
+                .reindex(expanded_range, fill_value=pad_value)
+            )
+        )
+        .fillna(pad_value)
+        .reset_index()
+    )
+    return df
+
+
+def prep_gps():
+    """prepare raw gps data"""
+    df = pd.read_csv(config.GPS_BB)
+    start_date = pd.to_datetime("2010-01-04")
+    end_date = pd.to_datetime("2011-01-02")
+    traj_len = "7d"
+    date_range = pd.date_range(
+        start=start_date,
+        end=end_date,
+        closed="left",
+        freq=traj_len,
+        name="DateRange",
+    )
+    x = pd.concat(
+        [
+            preprocess_time_gps(
+                df,
+                date_range[i],
+                date_range[i + 1],
+                sample_resolution="60min",
+                pad_value=0.0,
+            ).assign(i=i)
+            for i in range(len(date_range) - 1)
+        ]
+    )
+    return x
+
+
+def label_user_gaps(df, max_gap="60min"):
+    df["Gap"] = (df["DateTime"].diff() > max_gap).astype(int).cumsum()
+    return df
+
+
+def split_gaps(df, max_gap="60min"):
+    users = df.sort_values("DateTime").groupby("UID")
+    df = users.apply(label_user_gaps).reset_index()
+    return df
+
+
+def merge_datetime(df):
+    df["DateTime"] = pd.to_datetime(df["Date"] + " " + df["Time"])
+    df = df.drop(["Date", "Time"], axis=1)
+    return df
+
+
+def rowcol_to_1d(df):
+    df["Position"] = ((df["Row"] * df["Column"].max()) + df["Column"] + 1).astype(int)
+    df = df.drop(["Row", "Column"], axis=1)
+    return df
+
+
+def prep_data_sessions(maxlen_ntile=0.5):
+    """Prep mobility sequence session data split by session gaps"""
+    x = pd.read_csv(config.FM_MATRIX, index_col=0)
+    x = merge_datetime(x)
+    x = rowcol_to_1d(x)
+    x = split_gaps(x)
+    # get indexes
+    idxs = [
+        list(v.values)
+        for v in x.groupby(["UID", "Gap"]).groups.values()
+        if len(v.values) > 1
+    ]
+    x_sessions = [x.iloc[i] for i in idxs]
+    x_nested = [s.Position.tolist() for s in x_sessions]
+    # get 90th percentile sequence length
+    x_lengths = sorted([len(l) for l in x_nested])
+    pct_idx = round(len(x_lengths) * 0.5)
+    maxlen = x_lengths[pct_idx]
+    x_pad = pad_sequences(
+        x_nested, maxlen=maxlen, padding="pre", truncating="pre", value=0
+    )
+    x_enc = OrdinalEncoder().fit_transform(x_pad)
+    return x_enc
